Restructured folders of the project
