+++ web/src/components/Layout/AppBody/index.tsx
@@ -1,17 +1,17 @@
 import { useContext } from "react";
 import { css } from "@emotion/react";
 
-import { AuthContext } from "~/components/auth/AuthContext";
+import { AuthContext } from "~/context/AuthContext";
 import { useAppState } from "~/context/AppStateContext";
 import Panel from "~/components/Layout/Panel";
 
 import { Projects } from "~/components/Projects";
 
-import { RightMenu } from "./RightMenu";
-import { Search } from "./RightMenu/Search";
-import { TagsExplorer } from "./RightMenu/TagsExplorer";
+import { RightMenu } from "../RightMenu";
+import { Search } from "../RightMenu/Search";
+import { TagsExplorer } from "../RightMenu/TagsExplorer";
 import { Helmet } from "react-helmet-async";
-import Login from "~/components/auth/Login";
+import Login from "~/components/Auth/Login";
 
 const getCurrentPageName = (id: string, pages: any[]) => {
   return pages.find(proj => proj._id === id)?.name;
