+++ web/src/components/Modals/ConfirmDeletionModal.tsx
@@ -4,7 +4,7 @@ import { css } from "@emotion/react";
 
 import { THEME_COLORS } from "~/themeProvider";
 
-export interface ModalProps {
+interface ModalProps {
   onClose: () => void;
   onSubmit: (...args: any[]) => void;
 }
